Adjusting CLI Theme structure
