+++ src/Output/Theme/UnicornTheme.php
@@ -2,14 +2,14 @@
 
 namespace Minicli\Output\Theme;
 
-use Minicli\Output\CLITheme;
+
 use Minicli\Output\CLIColors;
 
-class UnicornTheme extends CLITheme
+class UnicornTheme extends DefaultTheme
 {
-    public function getDefaultColors()
+    public function getThemeColors(): array
     {
-        $styles = [
+        return [
             'default'     => [ CLIColors::$FG_CYAN ],
             'alt'         => [ CLIColors::$FG_BLACK, CLIColors::$BG_CYAN ],
             'error'       => [ CLIColors::$FG_RED ],
@@ -19,7 +19,5 @@ class UnicornTheme extends CLITheme
             'info'        => [ CLIColors::$FG_MAGENTA],
             'info_alt'    => [ CLIColors::$FG_WHITE, CLIColors::$BG_MAGENTA ]
         ];
-
-        return array_merge(parent::getDefaultColors(), $styles); // Any styles not defined here, will use the default styles.
     }
 }
